+++ apps/ai-dial-admin/src/components/PublicationView/PromptProperties/tests/PromptsProperties.spec.tsx
@@ -1,22 +1,38 @@
-import { render, screen } from '@testing-library/react';
-import { describe, it, expect } from 'vitest';
-import PromptsPropertiesList from '../PromptsPropertiesList';
+import { PublicationsI18nKey } from '@/src/constants/i18n';
+import { ActionType } from '@/src/models/dial/publications';
+import { fireEvent, render, screen } from '@testing-library/react';
+import { describe, expect, it } from 'vitest';
+import PromptsProperties from '../PromptsProperties';
 
-describe('PromptsPropertiesList', () => {
-  it('renders a list of PromptsProperties for each prompt', () => {
-    const publication = {
-      prompts: [{ name: 'Prompt1' }, { name: 'Prompt2' }],
-      action: 'UPDATE',
-      content: 'content',
-    };
-    render(<PromptsPropertiesList publication={publication as any} />);
-    expect(screen.getByText('Prompt1')).toBeInTheDocument();
-    expect(screen.getByText('Prompt2')).toBeInTheDocument();
+const basePrompt = {
+
+  name: 'PromptName',
+  version: '1.0',
+  description: 'desc',
+  content: 'Prompt content',
+};
+
+describe('PromptsProperties', () => {
+  it('renders collapsed by default', () => {
+    render(<PromptsProperties prompt={basePrompt} action={ActionType.ADD} collapsed={true} />);
+    expect(screen.getByText('PromptName')).toBeInTheDocument();
   });
 
-  it('renders nothing if no prompts', () => {
-    const publication = { prompts: [], action: 'UPDATE' };
-    const { container } = render(<PromptsPropertiesList publication={publication as any} />);
-    expect(container).toBeEmptyDOMElement();
+  it('expands and shows details when toggled', () => {
+    render(<PromptsProperties prompt={basePrompt} action={ActionType.ADD} collapsed={true} />);
+    fireEvent.click(screen.getByText('PromptName'));
+    expect(screen.getByText('desc')).toBeInTheDocument();
+    expect(screen.getByText('1.0')).toBeInTheDocument();
+    expect(screen.getByText('Prompt content')).toBeInTheDocument();
+  });
+
+  it('shows open button for DELETE action', () => {
+    render(<PromptsProperties prompt={basePrompt} action={ActionType.DELETE} collapsed={false} />);
+    expect(screen.getByRole('button', { name: PublicationsI18nKey.OpenPrompt })).toBeInTheDocument();
+  });
+
+  it('calls onOpenInNewTab when open button is clicked', () => {
+    render(<PromptsProperties prompt={basePrompt} action={ActionType.DELETE} collapsed={false} />);
+    fireEvent.click(screen.getByRole('button', { name: PublicationsI18nKey.OpenPrompt }));
   });
 });
